fix: no default profefe hostport for capture
The flags are colliding between commands. It means that the
profefe-hostport gets the default specified in order commands.

But for the kubectl profefe capture the empty string needs to be the
default one becuase it is used to discriminate if profiles needs to be
pushed in profefe or stored locally.

By default they get stored locally in /tmp

Signed-off-by: Gianluca Arbezzano <gianarb92@gmail.com>